app/ai-app/src/kdcube-ai-app/requirements-chat-ingress.txt
# Core API / gateway
aiohttp==3.12.15
fastapi==0.116.1
PyJWT[crypto]==2.10.1
pydantic==2.11.7
pydantic-settings==2.10.1
PyYAML==6.0.2
python-dotenv==1.1.1
python-socketio==5.13.0
redis==6.2.0
requests==2.32.5
uvicorn==0.37.0
asyncpg
psycopg2-binary==2.9.10
httpx
orjson
tenacity
aiolimiter

dramatiq==1.18.0

# Auth + storage (Cognito / S3)
-r requirements-aws.txt

# Uploads / AV / MIME preflight
python-multipart
aiofiles
python-magic
Pillow
clamd
beautifulsoup4

# Scheduling (opex routines)
croniter

# Link preview (disabled in ingress; Playwright only in proc)

# Payments (control plane)
stripe

tiktoken
langchain==1.2.0
langchain-community==0.4.1
langchain-openai==1.1.6
langchain-core==1.2.6
jsonschema

# Excel generation
openpyxl
# ------------------------------------------------------------
# Legacy bundle runtime + tools (keep if /integrations runs here)
# ------------------------------------------------------------

# langgraph==1.0.5
# langmem==0.0.30
# semantic-kernel
# ddgs
# mcp
# rich
# brotli
# markdown-it-py
# mdit-py-plugins
# linkify-it-py
# pygments
# jinja2
# markdownify
# trafilatura
# lxml
# Pillow
# python-dateutil
# networkx
# neo4j
# pandas
# matplotlib
# plotly
# kaleido
# graphviz
# seaborn
# diagrams
# pyvis
# tinycss2
# reportlab
# pymupdf
# pdfminer.six
# pypdf
# python-docx
# python-pptx
# xlsxwriter
# geopy
# folium
